# packages/@kit/api-client/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaaS Starter Kit API",
    "version": "1.0.0",
    "description": "A production-ready, type-safe REST API built with Hono and TypeScript.\n\n## Features\n- 🔐 JWT-based authentication with Better Auth\n- 👤 User management and profiles\n- 🔑 Role-based access control (RBAC)\n- ✉️ Email verification and password reset\n- 📝 Comprehensive OpenAPI documentation\n- 🎯 End-to-end type safety with Zod validation\n\n## Authentication\n\nMost endpoints require authentication via Bearer token:\n\n```\nAuthorization: Bearer YOUR_JWT_TOKEN\n```\n\nTo obtain a token:\n1. Sign up via POST /api/auth/sign-up/email\n2. Verify your email via the verification link\n3. Sign in via POST /api/auth/sign-in/email\n4. Use the returned session token in the Authorization header\n\n## Rate Limiting\n\nAPI requests are rate-limited to 100 requests per minute per IP address.",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "System",
      "description": "System health and status endpoints"
    },
    {
      "name": "User",
      "description": "User profile and account management"
    },
    {
      "name": "Admin",
      "description": "Administrative endpoints (requires admin role)"
    }
  ],
  "security": [
    {
      "Bearer": []
    }
  ],
  "components": {
    "schemas": {},
    "parameters": {}
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "healthCheck",
        "tags": [
          "System"
        ],
        "summary": "Health check",
        "description": "Check if the API is running and responding to requests.",
        "responses": {
          "200": {
            "description": "API is healthy and operational",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ],
                      "description": "Health status of the API"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp of the health check"
                    }
                  },
                  "required": [
                    "status",
                    "timestamp"
                  ]
                },
                "example": {
                  "status": "ok",
                  "timestamp": "2025-01-15T10:30:00.000Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": [
          "User"
        ],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's profile information including role and email verification status.\n\n**Use Cases:**\n- Display user profile in UI\n- Check user role for client-side authorization\n- Verify email verification status\n\n**Authentication:**\nRequires a valid Bearer token in the Authorization header.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Unique identifier for the user (UUID v4)"
                        },
                        "name": {
                          "type": "string",
                          "description": "User's full name"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "description": "User's email address (must be unique)"
                        },
                        "emailVerified": {
                          "type": "boolean",
                          "description": "Whether the user has verified their email address"
                        },
                        "image": {
                          "type": "string",
                          "nullable": true,
                          "description": "URL to the user's profile image (null if not set)"
                        },
                        "role": {
                          "type": "string",
                          "description": "User's role in the system (user, admin)"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "ISO 8601 timestamp of when the user was created"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "ISO 8601 timestamp of when the user was last updated"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "email",
                        "emailVerified",
                        "image",
                        "role",
                        "createdAt",
                        "updatedAt"
                      ],
                      "description": "Current authenticated user"
                    }
                  },
                  "required": [
                    "user"
                  ]
                },
                "example": {
                  "user": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "John Doe",
                    "email": "john.doe@example.com",
                    "emailVerified": true,
                    "image": "https://api.example.com/avatars/john.jpg",
                    "role": "user",
                    "createdAt": "2025-01-15T10:30:00.000Z",
                    "updatedAt": "2025-01-15T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required - No valid session or token provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    }
                  },
                  "required": [
                    "error",
                    "code"
                  ]
                },
                "example": {
                  "error": "Authentication required",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "operationId": "getAllUsers",
        "tags": [
          "Admin"
        ],
        "summary": "Get all users",
        "description": "Retrieve a list of all registered users in the system.\n\n**Requirements:**\n- Admin role required\n- Valid authentication token\n\n**Use Cases:**\n- User management dashboard\n- System administration\n- Analytics and reporting",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Unique identifier for the user (UUID v4)"
                          },
                          "name": {
                            "type": "string",
                            "description": "User's full name"
                          },
                          "email": {
                            "type": "string",
                            "format": "email",
                            "description": "User's email address (must be unique)"
                          },
                          "emailVerified": {
                            "type": "boolean",
                            "description": "Whether the user has verified their email address"
                          },
                          "image": {
                            "type": "string",
                            "nullable": true,
                            "description": "URL to the user's profile image (null if not set)"
                          },
                          "role": {
                            "type": "string",
                            "description": "User's role in the system (user, admin)"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "ISO 8601 timestamp of when the user was created"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "ISO 8601 timestamp of when the user was last updated"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "email",
                          "emailVerified",
                          "image",
                          "role",
                          "createdAt",
                          "updatedAt"
                        ]
                      },
                      "description": "Array of all users in the system"
                    }
                  },
                  "required": [
                    "users"
                  ]
                },
                "example": {
                  "users": [
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "John Doe",
                      "email": "john.doe@example.com",
                      "emailVerified": true,
                      "image": "https://api.example.com/avatars/john.jpg",
                      "role": "user",
                      "createdAt": "2025-01-15T10:30:00.000Z",
                      "updatedAt": "2025-01-15T10:30:00.000Z"
                    },
                    {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Jane Admin",
                      "email": "jane.admin@example.com",
                      "emailVerified": true,
                      "image": null,
                      "role": "admin",
                      "createdAt": "2025-01-10T08:15:00.000Z",
                      "updatedAt": "2025-01-14T16:20:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required - No valid session or token provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    }
                  },
                  "required": [
                    "error",
                    "code"
                  ]
                },
                "example": {
                  "error": "Authentication required",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin role required to access this endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "code": {
                      "type": "string",
                      "description": "Error code"
                    }
                  },
                  "required": [
                    "error",
                    "code"
                  ]
                },
                "example": {
                  "error": "Admin access required",
                  "code": "FORBIDDEN"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - Failed to fetch users from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "code": {
                      "type": "string",
                      "description": "Machine-readable error code"
                    },
                    "details": {
                      "nullable": true,
                      "description": "Additional error details"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Failed to fetch users",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    }
  }
}